src/servers/documentGrader/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-core
langchain-ollama
ipython